requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
cycler==0.10.0
dill==0.2.8.2
idna==2.7
inflection==0.3.1
kiwisolver==1.0.1
marshmallow==2.16.0
matplotlib==3.0.2
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
pillow-lut==1.0
pillow-scripts==5.0.0
pillowcase==2.0.0
pillowcover==0.0.6
pillowncase==0.3
pillowtalk==1.1.0a0
pycparser==2.19
pyparsing==2.3.0
pyreadline==2.1
python-dateutil==2.7.5
scipy==1.1.0
six==1.11.0
pytest==4.0.1
